fix gcc warning: pointer should be cast to unsigned long
tests/regression/01-cpa/29-fun_struct_array.c:15:24: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
   15 |   struct a A[1] = {50, (unsigned int)&QQ};
      |                        ^
In file included from tests/regression/01-cpa/29-fun_struct_array.c:2:
tests/regression/01-cpa/29-fun_struct_array.c:18:18: warning: comparison between pointer and integer
   18 |   assert(A[0].qq == &QQ); // UNKNOWN
      |                  ^~

+++ tests/regression/01-cpa/29-fun_struct_array.c
@@ -12,10 +12,10 @@ struct a{
 int main(){
   int i = 1;
 
-  struct a A[1] = {50, (int)&QQ};
+  struct a A[1] = {50, (unsigned long)&QQ};
 
   assert(A[0].aa == 50);
-  assert(A[0].qq == &QQ); // UNKNOWN
+  assert(A[0].qq == (unsigned long)&QQ); // UNKNOWN
 
   return 0;
 }
